Committing before refactoring.  NumberTranslatorTest fails.  Need to return the fractional part as an integer and round to the nearest hundredth if necessary (may not be necessary).

+++ src/test/java/com/tanabe/test/NumberTranslatorTest.java
@@ -2,8 +2,6 @@ package com.tanabe.test;
 
 import org.junit.Test;
 
-import java.math.BigDecimal;
-
 import static com.tanabe.rs.translator.NumberToEnglishTranslator.translateNumber;
 import static org.junit.Assert.assertEquals;
 
@@ -14,6 +12,6 @@ public class NumberTranslatorTest {
 
     @Test
     public void testTranslateTwoThousandFiveHundredTwentyThreeAndFourDollars(){
-        assertEquals("Two thousand five hundred twenty-three and 04/100 dollars", translateNumber(BigDecimal.valueOf(2523.04)));
+        assertEquals("Two thousand five hundred twenty-three and 04/100 dollars", translateNumber(2523.04));
     }
 }
